+++ src/main/java/dao/FichasClinicasHome.java
@@ -211,7 +211,7 @@ public class FichasClinicasHome {
         }
     }
 
-    public void delete(long id) {
+    public void delete(Integer id) {
         log.debug("deleting FichasClinicas instance");
         Transaction tx = null;
         Session session = sessionFactory.openSession();
